vehicle_stats/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
coverage>=7.3.2
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
django-cors-headers>=4.3.1
django-filter>=23.5 